# shared/openapi/openapi.yaml
openapi: 3.1.1
info:
  title: Szponcik communicator API
  version: 1.0.0
...
servers:
  - url: http://localhost:5000/api
    description: Local server
components:
  securitySchemes:
    TokenAuth:
      type: apiKey
      in: cookie
      name: access_token
    RefreshTokenAuth:
      type: apiKey
      in: cookie
      name: refresh_token
  schemas:
...
    Error:
      type: object
      required: [ error ]
      properties:
        error:
          type: string
          example: Something went wrong

    # ----- User -----
...
    IncomingWebhook:
      type: object
      required: [ id, channelId, name, url ]
      properties:
        id:
          type: string
        channelId:
          type: string
        name:
          type: string
        url:
          type: string
          description: Relative URL external services POST to in order to post a message.
        createdAt:
          type: string
    OutgoingWebhook:
      type: object
      required: [ id, name, url, format, events, isActive ]
      properties:
        id:
          type: string
        name:
          type: string
        url:
          type: string
        format:
          type: string
          enum: [ native, slack, discord ]
        events:
          type: array
          items:
            type: string
            enum: [ message.created, channel.created, member.added ]
        isActive:
          type: boolean
        secret:
          type: string
          description: HMAC signing secret (returned only when the webhook is created).
        createdAt:
          type: string
  parameters:
...
    WorkspaceID:
      name: workspaceId
      in: path
      description: The ID of the workspace.
      required: true
      schema:
        type: string
        format: uuid

    # ----- Channel -----
    ChannelID:
      name: channelId
      in: path
      description: The unique identifier of the channel.
      required: true
      schema:
        type: string
        format: uuid

    # ----- Message -----
...
  responses:
...
    '401':
      description: Unauthorized - Authentication is required and has failed or has not yet been provided
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    '403':
      description: Forbidden - The user does not have permission to access the resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    '404':
      description: Not Found - The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
security:
  - TokenAuth: [ ]
paths:
...
  /workspaces/{workspaceId}/channels/{channelId}/incoming-webhooks:
    post:
      operationId: createIncomingWebhook
      tags: [ Webhook ]
      summary: Create an incoming webhook for a channel
      description: Creates a token URL external services can POST to in order to post messages into the channel. Admins/owners only.
      security:
        - TokenAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/WorkspaceID'
        - $ref: '#/components/parameters/ChannelID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ name ]
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Incoming webhook created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncomingWebhook'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
  /workspaces/{workspaceId}/incoming-webhooks:
    get:
      operationId: listIncomingWebhooks
      tags: [ Webhook ]
      summary: List incoming webhooks
      security:
        - TokenAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/WorkspaceID'
      responses:
        '200':
          description: Incoming webhooks.
          content:
            application/json:
              schema:
                type: object
                required: [ webhooks ]
                properties:
                  webhooks:
                    type: array
                    items:
                      $ref: '#/components/schemas/IncomingWebhook'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
  /workspaces/{workspaceId}/incoming-webhooks/{webhookId}:
    delete:
      operationId: deleteIncomingWebhook
      tags: [ Webhook ]
      summary: Delete an incoming webhook
      security:
        - TokenAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/WorkspaceID'
        - name: webhookId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted.
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
  /workspaces/{workspaceId}/outgoing-webhooks:
    post:
      operationId: createOutgoingWebhook
      tags: [ Webhook ]
      summary: Create an outgoing webhook
      description: Registers a URL that receives a POST on the selected workspace events. Format native (signed JSON), slack or discord. Admins/owners only.
      security:
        - TokenAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/WorkspaceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ name, url, format, events ]
              properties:
                name:
                  type: string
                url:
                  type: string
                format:
                  type: string
                  enum: [ native, slack, discord ]
                events:
                  type: array
                  items:
                    type: string
                    enum: [ message.created, channel.created, member.added ]
      responses:
        '201':
          description: Outgoing webhook created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutgoingWebhook'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
    get:
      operationId: listOutgoingWebhooks
      tags: [ Webhook ]
      summary: List outgoing webhooks
      security:
        - TokenAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/WorkspaceID'
      responses:
        '200':
          description: Outgoing webhooks.
          content:
            application/json:
              schema:
                type: object
                required: [ webhooks ]
                properties:
                  webhooks:
                    type: array
                    items:
                      $ref: '#/components/schemas/OutgoingWebhook'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
  /workspaces/{workspaceId}/outgoing-webhooks/{webhookId}:
    delete:
      operationId: deleteOutgoingWebhook
      tags: [ Webhook ]
      summary: Delete an outgoing webhook
      security:
        - TokenAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/WorkspaceID'
        - name: webhookId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted.
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'